+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
             android:exported="false"
             android:label="Basic" />
         <activity
-            android:name=".conference.presentation.ConferenceActivity"
+            android:name=".feature_conference.presentation.ConferenceActivity"
             android:exported="false"
             android:label="conference" />
         <activity
