+++ app/src/main/java/com/survivalcoding/network_apps/MainActivity.kt
@@ -5,7 +5,7 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.ArrayAdapter
 import android.widget.ListView
-import com.survivalcoding.network_apps.conference.presentation.ConferenceActivity
+import com.survivalcoding.network_apps.feature_conference.presentation.ConferenceActivity
 import com.survivalcoding.network_apps.feature_basic.presentation.BasicActivity
 
 class MainActivity : AppCompatActivity() {
