--- app/src/main/java/com/survivalcoding/network_apps/conference/presentation/ConferenceState.kt
@@ -1,9 +1,0 @@
-package com.survivalcoding.network_apps.conference.presentation
-
-import com.survivalcoding.network_apps.conference.domain.model.ConferenceInfo
-
-// 변경사항
-data class ConferenceState(
-    val conferenceInfo: List<ConferenceInfo>? = null,
-    val isLoading: Boolean = false
-)
