+++ app/src/main/java/com/survivalcoding/network_apps/feature_conference/data/datasource/DataSource.kt
@@ -1,0 +1,8 @@
+package com.survivalcoding.network_apps.feature_conference.data.datasource
+
+import com.survivalcoding.network_apps.feature_conference.domain.model.ConferenceInfo
+
+// 변경사항
+interface DataSource {
+    suspend fun getConferenceList(): List<ConferenceInfo>
+}
