+++ app/src/main/java/com/survivalcoding/network_apps/feature_conference/domain/repository/ConferenceRepository.kt
@@ -1,0 +1,8 @@
+package com.survivalcoding.network_apps.feature_conference.domain.repository
+
+import com.survivalcoding.network_apps.feature_conference.domain.model.ConferenceInfo
+
+// 변경사항
+interface ConferenceRepository {
+    suspend fun getConferenceList(): List<ConferenceInfo>
+}
