+++ app/src/main/java/com/survivalcoding/network_apps/feature_conference/presentation/adapter/ConferenceDiffItemCallback.kt
@@ -1,7 +1,7 @@
-package com.survivalcoding.network_apps.conference.presentation.adapter
+package com.survivalcoding.network_apps.feature_conference.presentation.adapter
 
 import androidx.recyclerview.widget.DiffUtil
-import com.survivalcoding.network_apps.conference.domain.model.ConferenceInfo
+import com.survivalcoding.network_apps.feature_conference.domain.model.ConferenceInfo
 
 /*
 Note that DiffUtil, ListAdapter, and AsyncListDiffer require the list to not mutate while in use.
