+++ app/src/main/java/com/survivalcoding/network_apps/feature_conference/presentation/adapter/ConferenceViewHolder.kt
@@ -1,9 +1,9 @@
-package com.survivalcoding.network_apps.conference.presentation.adapter
+package com.survivalcoding.network_apps.feature_conference.presentation.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.survivalcoding.network_apps.conference.domain.model.ConferenceInfo
+import com.survivalcoding.network_apps.feature_conference.domain.model.ConferenceInfo
 import com.survivalcoding.network_apps.databinding.ItemConferenceBinding
 
 class ConferenceViewHolder(private val binding: ItemConferenceBinding) :
