+++ app/src/main/res/layout/activity_conference.xml
@@ -6,7 +6,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment_container_view"
-        android:name="com.survivalcoding.network_apps.conference.presentation.ConferenceFragment"
+        android:name="com.survivalcoding.network_apps.feature_conference.presentation.ConferenceFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
