+++ app/src/main/res/layout/fragment_conference.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".conference.presentation.ConferenceFragment">
+    tools:context=".feature_conference.presentation.ConferenceFragment">
